requirements.txt
streamlit
easyocr
opencv-python-headless
Pillow